+++ lib/src/models/controllers/view_controllers/schedule_view_controller.dart
@@ -91,12 +91,6 @@ abstract class ScheduleViewController extends ViewController with ScheduleMap {
     itemPositionsListener = list?.positionsListener;
   }
 
-  @Deprecated('Not used by the calendar. Use itemCountForPage. Will be removed in 0.33.0.')
-  int get itemCount => itemCountForPage(currentPage);
-
-  @Deprecated('Not used by the calendar. Use indexItem. Will be removed in 0.33.0.')
-  ListItem? item(int index) => indexItem(currentPage)[index];
-
   /// Get the [DateTime] for the given index of the current page.
   FloatingDateTime? dateTimeFromIndex(int index) => dateTimeFromIndexForPage(currentPage, index);
 
@@ -105,20 +99,6 @@ abstract class ScheduleViewController extends ViewController with ScheduleMap {
 
   /// Get the index closest to the given [DateTime] of the current page.
   int closestIndex(DateTime date) => closestIndexForPage(currentPage, date);
-
-  @Deprecated('Not used by the calendar. Use addItemForPage. Will be removed in 0.33.0.')
-  void addItem({required ListItem item, required FloatingDateTime date, bool isFirst = false}) {
-    return addItemForPage(item: item, date: date, pageIndex: currentPage, isFirst: isFirst);
-  }
-
-  @Deprecated('Not used by the calendar. Use clearPage. Will be removed in 0.33.0.')
-  void clear() => clearPage(currentPage);
-
-  @Deprecated('Not used by the calendar. Use closestIndex. Will be removed in 0.33.0.')
-  int initialScrollIndex(DateTime date) {
-    final normalized = FloatingDateTime.fromExternal(date, location: location).startOfDay;
-    return dateTimeItemIndex(currentPage)[normalized] ?? closestIndex(normalized);
-  }
 
   FutureOr<void> _animateToIndex(int index, {Duration? duration, Curve? curve}) {
     if (!hasInitialized) return null;